[refactor] MannerLevel을 InnerClass에서 Top-Level Class로 변경
