+++ src/main/java/com/prgrms/rg/domain/user/model/Manner.java
@@ -3,7 +3,6 @@ package com.prgrms.rg.domain.user.model;
 import static lombok.AccessLevel.*;
 
 import java.time.LocalDate;
-import java.util.Optional;
 
 import javax.persistence.Embeddable;
 
@@ -53,28 +52,5 @@ public class Manner {
 			"\n\tnoShow=" + noShow +
 			"\n\tbanned=" + bannedUntil +
 			"\n}";
-	}
-
-	public enum MannerLevel {
-		//TODO: 명칭 통일
-		NORMAL, FOUR_LEGS_BICYCLE, THREE_LEGS_BICYCLE, TWO_LEGS_BICYCLE, LEGENDARY_RIDER;
-
-		private static final short[] REQUIRED_FOR_PROMOTION = new short[] {500, 1000, 2000, 4000};
-
-		public static MannerLevel of(short point) {
-			if (point >= REQUIRED_FOR_PROMOTION[3]) {
-				return LEGENDARY_RIDER;
-			}
-			if (point >= REQUIRED_FOR_PROMOTION[2]) {
-				return TWO_LEGS_BICYCLE;
-			}
-			if (point >= REQUIRED_FOR_PROMOTION[1]) {
-				return THREE_LEGS_BICYCLE;
-			}
-			if (point >= REQUIRED_FOR_PROMOTION[0]) {
-				return FOUR_LEGS_BICYCLE;
-			}
-			return NORMAL;
-		}
 	}
 }
